Add more information about the current dev version (#231)
* Add more information about the current dev version

* Remove the GitHub commit link

This link will not always be accurate, for example if someone has made
local changes but not pushed them anywhere.  Showing a 404 in these
circumstances is not good.

As a future enhancement we can use a bit of JavaScript to query the
GitHub API and turn the commit into a link if it appears in our repo.