Remove contributors from readme
The 'contributors' list in the readme is disingenuous and misleading. I'm going to assume good intentions here and hope that you haven't done it on purpose, but listing the WordPress core team (and many other easily recognisable faces from the WordPress community) as your contributors is hugely misleading and makes it seem like this is an 'official' fork of WordPress.

You are a legitimate fork - don't mar that legitimacy by appearing to use shady tactics to make your project seem more significant. While it's nominally true that all of those folks have contributed to ClassicPress by virtue of their WordPress contributions, it is very obviously disingenuous to publish them in a list of contributors here.

I'm sure you can find a proper way of listing contributors here, but for now removing that list completely would be a very good idea.