Remove `$http_url` cruft
Follow-up to #201 and #142.

Removes variables that were set to `$http_url` in WordPress, because
WordPress API calls support HTTP without SSL (ClassicPress API calls do
not).

Leaves the existing "request retry" logic in place, and adds comments to
make it easier to find places where requests are retried in the future.