WP-r47178: Posts, Post Types: Fail gracefully when checking mapped cap against unregistered post status.
With `map_meta_cap` enabled for a post type, the `read_post` capability for posts with a public status is supposed to be mapped to the post type's `read` capability.

When a post is left in the database after the post status is no longer present, and WP does a `read_post` check against it, a PHP notice was thrown, and the cap check always failed.

As a more graceful fallback, the cap is now mapped onto `edit_others_posts`, which allows highly privileged users to be able to access orphaned content.

A `_doing_it_wrong()` notice is also added, so that developers and site administrators are aware that the cap mapping is failing in the absence of the registered post status.

Follow-up to https://core.trac.wordpress.org/changeset/34091, which introduced a similar approach to checking mapped caps against an unregistered post type.

WP:Props roytanck, SergeyBiryukov.
Fixes https://core.trac.wordpress.org/ticket/48653.

Conflicts:
  src/wp-includes/capabilities.php
  tests/phpunit/tests/user/capabilities.php
----
Merges https://core.trac.wordpress.org/changeset/47178 / WordPress/wordpress-develop@6f15251aa4 to ClassicPress.