Improve banner style
Before, the banner image was expected to be much wider than its height
(16x5 ratio or wider due to the 320x100 size enforced in the CSS). Now a
ratio of 4x3 or wider is allowed (320x240), with a ratio of 2x1 or 3x1
or higher being recommended.

To compensate for the increased height of a 2x1 image displayed in this
mode, the padding at the top of the login screen is also decreased.