Defer loading require()'d media JS files
Since rollup treats require() as a top-level import by default, this is
the easiest way to get rollup to execute code in the desired order.

I also tried the dynamicRequireTargets feature of rollup's commonjs
plugin but this did not generate working code.

This approach also has the benefit of not introducing any extra loader
code for module emulation.