Welcome new Contributors
I think, it will be of great value to add an automation to welcome new contributors like we do on the forums.
We would need to have a token to run the bot. 
It auto responds to the PR and also sends an email to the new contributor welcoming them.