WP-r48526: REST API: Issue a _doing_it_wrong when registering a route without a permission callback.
The REST API treats routes without a permission_callback as public. Because this happens without any warning to the user, if the permission callback is unintentionally omitted or misspelled, the endpoint can end up being available to the public. Such a scenario has happened multiple times in the wild, and the results can be catostrophic when it occurs.

For REST API routes that are intended to be public, it is recommended to set the permission callback to the `__return_true` built in function.

Fixes https://core.trac.wordpress.org/ticket/50075.
WP:Props rmccue, sorenbronsted, whyisjake, SergeyBiryukov, TimothyBlynJacobs.

Conflicts:
- src/wp-includes/class-wp-oembed-controller.php
- src/wp-includes/rest-api/endpoints/class-wp-rest-post-types-controller.php
- src/wp-includes/rest-api/endpoints/class-wp-rest-users-controller.php
- tests/phpunit/tests/rest-api.php
- tests/phpunit/tests/rest-api/rest-server.php

---

Merges https://core.trac.wordpress.org/changeset/48526 / WordPress/wordpress-develop@74cc64d74e to ClassicPress.