Multiple select support for wp_dropdown_categories
Adds a new argument to the $args array for wp_dropdown_categories, allowing to pass either empty string or `multiple`. 

If left empty, the select defaults to a single select, if `multiple` is passed the Select will allow for multiple selection and pass an `[]` to the name attribute value as well.

This was tested on several installs of mine (I use to replace said function with a custom one). Also tested this particular commit on my local install
Example usage:
`wp_dropdown_categories( array('multiple' => 'multiple') );`