WP-r46175: Improve `do_enclose()` logic on post publish.
Removing the direct SQL query in `do_all_pings()` improves filterability.

As part of this change, the signature of `do_enclose()` is changed so that
a null `$content` parameter can be passed, with the `$content` then inferred
from the `$post` passed in the second parameter. In addition, the second
parameter was modified so that a post ID or a `WP_Post` object can be
provided. These changes make it possible to trigger enclosure checks with
a post ID alone (as in `do_all_pings()`) and also brings the function
signature in line with `do_trackbacks()` and `pingback()`.

WP:Props dshanske, spacedmonkey, janw.oostendorp, mrmadhat, birgire.
See https://core.trac.wordpress.org/ticket/36824.

Conflicts:
- src/wp-includes/comment.php
- src/wp-includes/functions.php

---

Merges https://core.trac.wordpress.org/changeset/46175 / WordPress/wordpress-develop@48e280db32 to ClassicPress.