Enable comment form enctype to be changed
This PR adds an action within the comment form tag,. This makes it possible to change the form to have  the attribute `enctype="multipart/form-data"`, which enables uploads with the comment. The method used here mirrors that already used for posts and categories.

Without this PR, the only way to achieve this is with JavaScript.

The PR also corrects a misleading statement about the `do_action( 'comment_form_top' )` that follows . This currently states that it fires inside the form tag but, in fact, it fires just after the form tag and so is useless for changing the form's `enctype`.