Listen to #submenu presence instead of $menu
Instead of checking if $menu global has a key 85 and said array key has a second array member of value `security.php`, we simply check if the $submenu global has a key of value `security.php`, in which case it means plugin added add_security_page on admin_menu hook

We remove the menu if no such key is found. This still grants the menu shown when plugin adds it. And removes it elsewise/