Add support for wp_date() function
WP 5.3 has introduced a new, non-legacy-dependent Date Function `wp_date`.
Several _important_ plugins such as Postmark rely on it. This means, you can't use Postmark on CP - Postmark is the _leading_ solution for Mail Sending on VPS servers that do not come with inbuilt mailers.

This PR simply adds support for `wp_date`. It does NOT change the core code to _use_ that new function, but, if any plugin relies on it, it will be there, thus avoiding fatal errors.

To support `wp_date` 2 more functions have to be added (included in this PR):
- `wp_timezone`
- `wp_timezone_string`

This PR is not a classic back port, because it does _not_ make Core use these new functions, it just adds support for them, which is all we need to have Plugins (and other code) working with CP.