WP-r53900: Mail: Prevent the last character of names in  “From” headers from being trimmed.
When extracting the email and name from a “From” header, the last character of the name is incorrectly trimmed when a space is not included between the name and the opening `<`.

Though the space is required for the header to be compliant with RFC5322 (see https://www.rfc-editor.org/rfc/rfc5322#section-3.4), the absence of a space can be ignored here. PHPMailer accepts the name and email as separate parameters and constructs the header correctly later on.

WP:Props hakanca, mikehansenme, SergeyBiryukov, kovshenin, mattyrob, drewapicture, desrosj.
Fixes https://core.trac.wordpress.org/ticket/19847.

Conflicts:
- tests/phpunit/tests/mail.php

---

Merges https://core.trac.wordpress.org/changeset/53900 / WordPress/wordpress-develop@002ca62974 to ClassicPress.