WP-r42832: General: Introduce dashboard widget to inform administrators about outdated PHP versions.
This new dashboard widget is shown on WordPress sites which are powered by a PHP version which WordPress considers outdated, in order to inform site owners about the resulting problems and to explain how to upgrade to a supported version. An education page for that purpose has been previously created that the widget links to. The link is translatable so that localized versions of the page can be referred to as they become available.

The nag follows the example of the Browse Happy dashboard widget and is only visible for administrators, or network administrators when using multisite. To determine whether it needs to be displayed, a new wordpress.org API introduced prior is called that handles the version logic in a centralized location.

WP:Props flixos90, hedgefield, schlessera.
Fixes https://core.trac.wordpress.org/ticket/41191.

Conflicts:
- src/wp-admin/css/dashboard.css
- src/wp-includes/capabilities.php
- tests/phpunit/tests/user/capabilities.php

---

Merges https://core.trac.wordpress.org/changeset/42832 / WordPress/wordpress-develop@5f56921131 to ClassicPress.