Fix: Remove Accordion script and update Nav Menus (#205)
* Update template.php

Following PR #197, there is no longer any need for the accordion script. This PR therefore removes it from the JS folder in `wp-admin` and removes the function calling it from `template.php` in `wp-admin/includes`

* Remove jQuery accordion script from those registered in script-loader.php

* Update script-loader.php

Remove more references to the accordion script

* Reverting earlier commit because it targeted the wrong 'accordion' script

Confusingly, there are two accordion scripts in the CP admin: one is part of jQuery UI and should not be removed. This commit reverts its removal.

* Deatils and summary tags added to class-wp-customize-nav-menus.php

This replicates the changes made to the regular `nav menus` screen in that it adds the `details` and `summary` tags to the Add Menus section in the Customizer.

This change both improves accessibility and enables the `accordion` script to be removed.

* Update common.css

This deletes `display: none;` from the `.accordion-section-content` class and so enables the relevant sections to become visible when toggled using the `details` tag.

* Coding standards and docblock corrections

* Fix for failing unit test

* Delete src/wp-admin/js/accordion.js

Delete file as it is no longer used

---------

Co-authored-by: mattyrob <mattyrobuk@gmail.com>