Update post metaboxes and dashboard widgets with modern HTML (#217)
* Adds details and summary elements to post metaboxes and dashboard widgets

This enhances accessibility for post metaboxes and dashboard widgets and enables them to be toggled open and closed without JavaScript.

* Update edit.css

Ensures that post metaboxes and dashboard widgets look the same as before the use of details and summary elements.

* Update common.css

Enures that post metaboxes and dashboard widgets look the same as before the use of details and summary elements.

* Update postbox.js

Ensures that open/closed state of post metaboxes and dashboard widgets is preserved across page loads. Before it was based on the presence or otherwise of the class "closed"; now it is dependent on the presence or otherwise of the attribute "open".

* Update sortable.js

Sortable widgets in the CP backend currently rely on jQuery UI. While it would be nice to replace that with something more modern (like SortableJS), that has to wait for another day. In the meantime, we need a fix to address the fact that the HTML5 elements `details` and `summary` were introduced after development ceased on jQuery UI.

This fix simply avoids the presence of an ugly artefact when dragging and dropping a `details` element.

* WP-r56622: Add `sys_get_temp_dir()` to `open_basedir` tests. (#218)

In PHPUnit 10.3.5, 9.6.13 and 8.5.34, the child processes used for process isolation now use temporary files to communicate their result to the parent process.

This caused a failure in some tests that set the `open_basedir` PHP directive to a value that did not include `sys_get_temp_dir()`.

This adds `sys_get_temp_dir()` to the `open_basedir` value set by the tests to ensure that permission is still granted for the temporary directory.

PHPUnit uses `sys_get_temp_dir()`. To ensure the result is the same, Core's `get_temp_dir()` function is not used.

References:
- https://github.com/sebastianbergmann/phpunit/issues/5356

WP:Props desrosj, mukesh27, SergeyBiryukov, costdev.
Fixes https://core.trac.wordpress.org/ticket/59394.

---

Merges https://core.trac.wordpress.org/changeset/56622 / WordPress/wordpress-develop@7d96189ba1 to ClassicPress.

Co-authored-by: Colin Stewart <costdev@git.wordpress.org>

* Renovate[bot]: Update dependency node to v18.18.0 (#216)

Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>

* Adds details and summary elements to post metaboxes and dashboard widgets

This enhances accessibility for post metaboxes and dashboard widgets and enables them to be toggled open and closed without JavaScript.

* Update edit.css

Ensures that post metaboxes and dashboard widgets look the same as before the use of details and summary elements.

* Update common.css

Enures that post metaboxes and dashboard widgets look the same as before the use of details and summary elements.

* Update postbox.js

Ensures that open/closed state of post metaboxes and dashboard widgets is preserved across page loads. Before it was based on the presence or otherwise of the class "closed"; now it is dependent on the presence or otherwise of the attribute "open".

* Update sortable.js

Sortable widgets in the CP backend currently rely on jQuery UI. While it would be nice to replace that with something more modern (like SortableJS), that has to wait for another day. In the meantime, we need a fix to address the fact that the HTML5 elements `details` and `summary` were introduced after development ceased on jQuery UI.

This fix simply avoids the presence of an ugly artefact when dragging and dropping a `details` element.

* PHP Coding standards

* JS Coding Standards

* Update common.js to address Safari's failure to respect effect of closure of details element

Addresses the focusable issue mentioned in this comment: https://github.com/ClassicPress/ClassicPress-v2/pull/217#issuecomment-1735439355

This commit applies the `inert` attribute to elements within a closed `details` element. This appears to be supported by both desktop and mobile versions of Safari: see https://caniuse.com/?search=inert

* Update common.css to replace details arrows via CSS so that they can also have a pointer cursor.

* Update common.css to improve style of arrows

* JS syntax corrections

* Update common.css to make `details` marker bigger and remove the default in Safari.

* Compliant CSS comment

* Suggested Safari fix

* Update common.css for summary header and custom marker

* Update template.php to add back `h2` and `h3` tags.

This responds to feedback at https://github.com/ClassicPress/ClassicPress-v2/pull/217#discussion_r1341996477

This should also resolve the issue where clicking on the pointer icon still enabled the widget to be dragged. Dragging is now possible only where the draggable icon appears.

* Update template.php to move a class to avoid causing problems on the admin menus page.

* Update common.css to modify styling on admin menus page.

The changes made elsewhere in the PR have an impact on the visual styling of the admin menus page. For reasons long forgotten, and as noted in this file already, that page works differently from other admin pages that use widgets. So some modifications are necessary to ensure the menus page continues to look as expected.

* Update template.php to remove tabindex and enqueuing of deleted script.

* Fix whitespace in wp-admin/includes/template.php

---------

Co-authored-by: Matt Robinson <mattyrobuk@gmail.com>
Co-authored-by: Colin Stewart <costdev@git.wordpress.org>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>