Fix Restore Tinymce editor in text widget (#275)
* Update script-loader.php to include scripts required to run TinyMCE in the text widget

See Issue #273

* Update text-widgets.js to run TinyMCE editor

* Update script-loader.php to remove unrelated code

* Update text-widgets.js to add TinyMCE bypasses

* Update text-widgets.js

This uses a combination of vanilla JS and jQuery, but that's what the widget uses already. (Yes, the code we've inherited here is a real dog's breakfast.) At a later date we can work on replacing this with purely vanilla JS once we have it all working.

This still leaves one thing not working correctly. If two Text widgets are added in the Customizer, the second won't open; that probably needs a fix in a different file. I will investigate.

* Coding standards fix

* Update script-loader.php in response to https://github.com/ClassicPress/ClassicPress-v2/pull/275#pullrequestreview-1708604020

* Update text-widgets.js

* Update customize-widgets.js to make TinyMCE work for all Text widgets in Customizer

* Update text-widgets.js to delay instantiating TinyMCE for 200ms

* Update widgets.css to prevent widgets expanding beyond container

---------

Co-authored-by: mattyrob <mattyrobuk@gmail.com>