Reinstate `jquery-ui-widget` as its own file
This file was absorbed into `jquery-ui-core` in jQuery UI v 1.12.1. The PR takes it back out into its own file once again (but this time without a dependency on `jquery-ui-core` or, indeed, on any other file) because it's the only file from jQuery UI that is worth preserving. This will mean that it will be treated like a library and simply become a specific dependency for `iris.js` and for nothing else, while its presence (and consequent modifications to `wp-includes/script-loader.php`) will preserve backward compatibility.