WP-r53557: Code Modernization: Remove dynamic properties in `Tests_*_Slashes`.
Dynamic (non-explicitly declared) properties are deprecated as of PHP 8.2 and are expected to become a fatal error in PHP 9.0.

In this particular case, the test classes contain a `set_up()` method that sets a group of properties, which are ''used'' by the tests, but the values of these properties are never ''changed'' by the tests.

In other words, setting these properties in the `set_up()` is an unnecessary overhead and the properties should be changed to class constants.

Follow-up to [1041/tests], [1071/tests].

WP:Props jrf.
See https://core.trac.wordpress.org/ticket/56033.

Conflicts:
- tests/phpunit/tests/post/slashes.php
- tests/phpunit/tests/user/slashes.php

---

Merges https://core.trac.wordpress.org/changeset/53557 / WordPress/wordpress-develop@bafecbeab5 to ClassicPress.