Update media.php to move function
This PR removes the function `cp_select_upload_media_category()` from its current position and instead hooks it to the `pre-upload-ui` action hook. This is the hook plugins have used to modify the attachments upload behavior.

Note that this does not solve the issues reported above with this PR. What it does instead is ensure that all browsers respond consistently, and so provides a better basis on which to address those issues.