WP-r59108: Code Modernization: Fix trigger_error() with E_USER_ERROR deprecation in WP_Test_Stream::open().
PHP 8.4 deprecates the use of `trigger_errror()` with `E_USER_ERROR` as the error level, as there are a number of gotchas to this way of creating a `Fatal Error` (`finally` blocks not executing, destructors not executing).
The recommended replacements are either to use exceptions or to do a hard `exit`.

As this is a test-only class, do not have to take BC-breaks into account.

Also, as this is a test helper, throwing a exception is the most appropriate solution.

Reference:
* https://wiki.php.net/rfc/deprecations_php_8_4#deprecate_passing_e_user_error_to_trigger_error

Follow-up to https://core.trac.wordpress.org/changeset/49230.

WP:Props jrf.
See https://core.trac.wordpress.org/ticket/62061.

---

Merges https://core.trac.wordpress.org/changeset/59108 / WordPress/wordpress-develop@004379e56b to ClassicPress.