Add files via upload
Updated file functions.php. 
Removed several functions that are added for the main CP site or that are not relevant / not necessary. 
Moved 2 functions ("Add the page slug as a class to the <body>" and "Add excerpts to pages") to file template-functions.php. 
Included custom header (file inc/custom-header.php becomes redundant now).