Add files via upload
Updated file functions.php. 
Removed functions that are added for the main CP site or that are not relevant / not necessary. 
- susty_nav_rewrite_rule()
- susty_register_query_var()
- cp_susty_get_asset_version()
- filter_featured_image_admin_text()
- cp_remove_empty_p()
- is_blog()
- cp_susty_override_style_css_version()
Moved 2 functions ("Add the page slug as a class to the <body>" and "Add excerpts to pages") to file template-functions.php. 
Included custom header (file inc/custom-header.php becomes redundant now).