Update class-wp-widget-media-gallery.php to enqueue jQuery UI Sortable module
This is a temporary fix to ensure that the images are sortable within the widget until all remnants of backbone.js are removed from widgets. Enqueuing the script from here means that it does not have to be enqueued from the widgets page and so does not produce a deprecation warning when the page loads.