Refactor Media gallery widget to vanilla JavaScript and PHP (#1820)
* Update media-gallery-widget.js to avoid using backbone.js to build widget

* Update class-wp-widget-media-gallery.php to rebuild using PHP instead of backbone.js

* Update widgets.css

* Update media-gallery-widget.js to localize a label

* Update class-wp-widget-media-gallery.php to localize a label

* Update class-wp-widget-media-gallery.php to enqueue jQuery UI Sortable module

This is a temporary fix to ensure that the images are sortable within the widget until all remnants of backbone.js are removed from widgets. Enqueuing the script from here means that it does not have to be enqueued from the widgets page and so does not produce a deprecation warning when the page loads.

* Update widgets-form.php to no longer enqueue jQuery UI Sortable module.

The script was enqueued from here as a temporary fix in PR #1713 to ensure that images in the gallery widget could still be sorted. The script will now be enqueued from the widget itself until the script is replaced with Sortable.js.

* Update class-wp-widget-media-gallery.php to restore required function

* Update class-wp-widget-media-gallery.php

* Update media-gallery-widget.js

* Update class-wp-widget-media-gallery.php to comply with coding standards

* Update media-gallery-widget.js to eliminate unused variable

* Remove unnecessary QUnit tests

* Remove unnecessary PHPUnit test

* Remove unnecessary PHPUnit assertion

* Use parent class `update()` function for sanitization

* Update class-wp-widget-media-gallery.php to use a string instead of an array within the widget form

* Update media-gallery-widget.js to correct indentation

* Update class-wp-widget-media-gallery.php to remove enqueuing of jQuery Sortable script

Enqueuing jQuery Sortable script rendered unnecessary by PR #1832

---------

Co-authored-by: mattyrob <mattyrobuk@gmail.com>

CP:Props KTS915, mattyrob