WP-r55938: Filesystem API: Attempt to create directory in `copy_dir()`.
Adds a check to the start of `copy_dir()` that the destination directory exists and attempts to create it if it does not.

An error is returned if the directory can not be created, either due to a permissions error or the parent directory not existing.

WP:Props caraffande, costdev, zunaid321.
Fixes https://core.trac.wordpress.org/ticket/41855.

---

Merges https://core.trac.wordpress.org/changeset/55938 / WordPress/wordpress-develop@7cfc7a46f9 to ClassicPress.